pyproject.toml
[project]
name = "voice2text"
version = "0.1.0"
description = "語音辨識發送文字系統 - Qwen3-ASR + OpenCC + Translation + History"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Voice2Text Team" }
]
keywords = ["asr", "speech-recognition", "qwen", "opencc", "translation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "qwen-asr>=0.1.0",
    "sounddevice>=0.4.0",
    "scipy>=1.11.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PySide6>=6.5.0",
    "typer>=0.9.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "opencc-python-reimplemented>=0.1.7",
    "aiosqlite>=0.19.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
voice2text = "app.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
